# openapi.yml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Dushuli
  description: "读书历CRUD服务"
  license:
    name: MIT
servers:
  - url: https://api.36node.com{basePath}
    description: The production API server
    variables:
      basePath:
        default: /dushuli/v0
paths:
  /wechat/payment:
    post:
      summary: initiate a payment process
      operationId: createPayment
      tags:
        - wechat
      requestBody:
        description: payment info
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentBody"
      responses:
        "200":
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatePaymentResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /wechat/signature:
    get:
      summary: get wechat js sdk signature
      operationId: getSignature
      tags:
        - wechat
      parameters:
        - name: url
          in: query
          description: 调用 jssdk 的网页链接
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetSignatureResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /products:
    get:
      summary: list all products
      operationId: listProducts
      tags:
        - product
      parameters:
        - name: _select
          in: query
          description: select which fields from db
          required: false
          schema:
            type: string
        - name: _limit
          in: query
          description: how many items to return at one time (max 100)
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: _offset
          in: query
          description: How many items will be skip
          required: false
          schema:
            type: string
        - name: _sort
          in: query
          description: specifies which document fields to sort
          required: false
          schema:
            type: string
        - name: published
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: a paged array of products
          headers:
            X-Total-Count:
              description: total count of all products match the query
              schema:
                type: string
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
    
    post:
      summary: create a product
      operationId: createProduct
      tags:
        - product
      requestBody:
        description: product to be created
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product" 
      responses:
        "201":
          description: the product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /products/{productId}:
    get:
      summary: get product by productId
      operationId: getProduct
      tags:
        - product
      parameters:
        - name: productId
          in: path
          required: true
          description: the productId of the product to retrieve
          schema:
            type: string
      responses:
        "200":
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
        
    delete:
      summary: delete a product
      operationId: deleteProduct
      tags:
        - product
      parameters:
        - name: productId
          in: path
          description: the productId of the product to be deleted
          required: true
          schema:
            type: string
      responses:
        "204":
          description: product deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    put:
      summary: update product
      operationId: updateProduct
      tags:
        - product
      parameters:
        - name: productId
          in: path
          description: the user of the product to be updated
          required: true
          schema:
            type: string
      requestBody:
        description: product to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product"
      responses:
        "200":
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"    
...
security:
  - bearerAuth: []
components:
  responses:
...
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Err"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Err"
    InternalError:
      description: unexpected error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Err"
  schemas:
    CreatePaymentBody:
      type: object
      required:
        - openid
      properties:
        openid:
          type: string
          description: 付款用户的openid
        sandbox:
          type: boolean
          description: 是否是沙盒模式
    CreatePaymentResponse:
      type: object
      properties:
        appid:
          type: string
          description: 公众号 appid
          example: "wx8888888888888888"
        timeStamp: 
          type: string
          description: 时间戳
          example: "1414561699"
        nonceStr:
          type: string
          description: 随机字符串
          example: "5K8264ILTKCH16CQ2502SI8ZNMTM67VS"
        package:
          type: string
          description: 订单详情扩展字符串
          example: "prepay_id=123456789"
        signType:
          type: string
          description: 签名方式
          example: "MD5"
        paySign:
          type: string
          description: 签名
          example: "C380BEC2BFD727A4B6845133519F3AD6"
    GetSignatureResponse:
      type: object
      properties:
        debug:
          type: boolean
        appId:
          type: string
          description: 公众号 appid
          example: "wx8888888888888888"
        timestamp:
          type: string
          description: 时间戳
          example: "1414561699"
        nonceStr:
          type: string
          description: 随机字符串
          example: "5K8264ILTKCH16CQ2502SI8ZNMTM67VS"
        signature:
          type: string
          description: 签名
          example: "5bfc8633f9c2127508296408c64f99068ad9dcb5"
        jsApiList:
          type: array
          items:
            type: string
          description: js api 列表
          example: [
            "checkJsApi",
            "onMenuShareTimeline",
            "onMenuShareAppMessage",
            "onMenuShareQQ",
            "onMenuShareWeibo",
            "onMenuShareQZone",
            "chooseImage",
            "previewImage",
            "uploadImage",
            "startRecord",
            "stopRecord",
            "onVoiceRecordEnd",
            "playVoice",
            "pauseVoice",
            "stopVoice",
            "onVoicePlayEnd",
            "uploadVoice",
            "downloadVoice",
            "translateVoice",
            "getNetworkType",
            "openLocation",
            "getLocation",
            "startSearchBeacons",
            "stopSearchBeacons",
            "onSearchBeacons",
            "closeWindow",
            "hideMenuItems",
            "showMenuItems",
            "hideAllNonBaseMenuItem",
            "showAllNonBaseMenuItem",
            "scanQRCode",
            "openProductSpecificView",
            "chooseCard",
            "addCard",
            "openCard",
            "chooseWXPay"
          ]
...
    Product:
      type: object
      properties:
        id:
          type: string
          description: id
        createdAt:
          type: string
          format: date-time
          description: 创建时间
        updatedAt:
          type: string
          format: date-time
          description: 更新时间
        slug:
          type: string
          description: slug
        price:
          type: number
          description: 价格
        name:
          type: string
          description: 商品名称
        description:
          type: string
          description: 商品描述
        start:
          type: string
          format: date-time
          description: 有效期起始时间
        end:
          type: string
          format: date-time
          description: 有效期结束时间
        period:
          type: number
          description: 延长有效期（毫秒）
        published:
          type: boolean
          description: 是否发布
        publishedAt:
          type: string
          format: date-time
          description: 发布时间
    Err:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT